refactor(core): extract signal cell base classes, shed ngrx dataSource
