+++ src/frontend/packages/core/src/features/endpoints/backup-restore/backup-checkbox-cell/backup-checkbox-cell.component.ts
@@ -3,7 +3,7 @@ import { FormsModule } from '@angular/forms';
 import { CustomCheckboxComponent } from '../../../../shared/components/custom-checkbox/custom-checkbox.component';
 import { EndpointModel } from '@stratosui/store';
 
-import { TableCellCustom } from '../../../../shared/components/list/list.types';
+import { TableCellCustom } from '../../../../shared/components/signal-list/cell-base';
 import { BackupEndpointsService } from '../backup-endpoints.service';
 import { BackupEndpointTypes } from '../backup-restore.types';
